refactor timer support into infra layer and update generic timer access
  - split timer infrastructure helpers into acs_timer_infra
  - move generic timer register APIs into acs_timer.h
  - add EL2/VHE-aware timer register read/write handling
  - program EL1/EL2 timers through CVAL instead of TVAL
  - update build files, headers, and timer/ETE call sites to use the new interfaces
