+++ test_pool/ete/ete004.c
@@ -17,7 +17,7 @@
 
 #include "acs_val.h"
 #include "acs_pe.h"
-#include "acs_timer_support.h"
+#include "acs_timer.h"
 #include "acs_ete.h"
 #include "val_interface.h"
 #include "acs_memory.h"
