+++ test_pool/timer/t001.c
@@ -16,7 +16,7 @@
  **/
 
 #include "acs_val.h"
-#include "acs_timer.h"
+#include "acs_timer_infra.h"
 
 #define TEST_NUM   (ACS_TIMER_TEST_NUM_BASE + 1)
 #define TEST_RULE  "B_TIME_01"
