+++ val/include/val_sysreg_timer.h
@@ -17,6 +17,7 @@ SYSREG_RW_FUNCS(cnthp_tval_el2)
 SYSREG_RW_FUNCS(cnthv_ctl_el2)
 SYSREG_RW_FUNCS(cnthv_tval_el2)
 SYSREG_RW_FUNCS(cnthp_cval_el2)
+SYSREG_RW_FUNCS(cnthv_cval_el2)
 SYSREG_RW_FUNCS(cntps_ctl_el1)
 SYSREG_RW_FUNCS(cntps_tval_el1)
 SYSREG_RW_FUNCS(cntps_cval_el1)
